api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pydantic==2.10.6
python-dotenv==1.0.1
